refactor(transfer): decompose delta_pipeline.rs into stage submodules (SPL-20 #2447)
