+++ crates/transfer/src/delta_pipeline/tests/mod.rs
@@ -1,0 +1,4 @@
+mod helpers;
+mod parallel;
+mod sequential;
+mod threshold;
